// src/store/auth/types.ts
export enum AuthActionTypes {
  FETCH_AUTH_REQUEST = "@@AUTH/FETCH/AUTH_REQUEST",
  FETCH_AUTH_ERROR = "@@AUTH/FETCH/AUTH_ERROR",
  FETCH_AUTH_SUCCESS = "@@AUTH/FETCH/AUTH_SUCCESS",
  VALIDATION = "@@AUTH/VALIDATION",
  CLEAR_REQUEST = "@@AUTH/CLEAR/REQUEST",
  CLEAR = "@@AUTH/CLEAR",
  DETAIL_USER_SUCCESS = "@@AUTH/FETCH_DETAIL_USER_SUCCESS"
}